requirements.txt
thordata-sdk>=1.7.0
python-dotenv
pandas
colorama
requests
